refactor(seed): move test appointment seeding logic to a separate file for better organization
